refactored plaid-related api routes
