+++ server/index.js
@@ -20,11 +20,11 @@ const store = new MongoDBStore({
   collection: 'mySessions',
 });
 
-const client = new plaid.Client({
-  clientID: process.env.PLAID_CLIENT_ID,
-  secret: process.env.PLAID_SECRET,
-  env: plaid.environments.sandbox,
-});
+// const client = new plaid.Client({
+//   clientID: process.env.PLAID_CLIENT_ID,
+//   secret: process.env.PLAID_SECRET,
+//   env: plaid.environments.sandbox,
+// });
 
 store.on('error', function (error) {
   console.log(error);
@@ -59,92 +59,84 @@ app.use(
 app.use(passport.initialize());
 app.use(passport.session());
 
-app.get('/link/token/create', async (req, res) => {
-  try {
-    const { link_token } = await client.createLinkToken({
-      user: {
-        client_user_id: '123-test-user-id',
-      },
-      client_name: 'Plaid Test App',
-      products: ['auth', 'transactions'],
-      country_codes: ['US'],
-      language: 'en',
-      // webhook: 'https://sample-web-hook.com',
-      account_filters: {
-        depository: {
-          account_subtypes: ['checking', 'savings'],
-        },
-      },
-    });
-    res.send(link_token);
-  } catch (error) {
-    console.error(error);
-  }
-});
+// app.get('/link/token/create', async (req, res) => {
+//   try {
+//     const { link_token } = await client.createLinkToken({
+//       user: {
+//         client_user_id: '123-test-user-id',
+//       },
+//       client_name: 'Plaid Test App',
+//       products: ['auth', 'transactions'],
+//       country_codes: ['US'],
+//       language: 'en',
+//       // webhook: 'https://sample-web-hook.com',
+//       account_filters: {
+//         depository: {
+//           account_subtypes: ['checking', 'savings'],
+//         },
+//       },
+//     });
+//     res.send(link_token);
+//   } catch (error) {
+//     console.error(error);
+//   }
+// });
 
-app.post('/plaidTokenExchange', async (req, res) => {
-  try {
-    const user = await db.collection('users').findOne({ _id: req.user._id });
-    const { publicToken } = req.body;
+// app.post('/plaidTokenExchange', async (req, res) => {
+//   try {
+//     const user = await db.collection('users').findOne({ _id: req.user._id });
+//     const { publicToken } = req.body;
 
-    if (user.access_token.length) {
-      const { accounts, item } = await client
-        .getAccounts(user.access_token)
-        .catch(handleError);
-      res.send(user.access_token);
-    } else {
-      const { access_token } = await client
-        .exchangePublicToken(publicToken)
-        .catch(handleError);
-      const { accounts, item } = await client
-        .getAccounts(access_token)
-        .catch(handleError);
+//     if (user.access_token.length) {
+//       const { accounts, item } = await client
+//         .getAccounts(user.access_token)
+//         .catch(handleError);
+//       res.send(user.access_token);
+//     } else {
+//       const { access_token } = await client
+//         .exchangePublicToken(publicToken)
+//         .catch(handleError);
+//       const { accounts, item } = await client
+//         .getAccounts(access_token)
+//         .catch(handleError);
 
-      db.collection('users').updateOne(
-        { _id: req.user._id },
-        { $set: { access_token } }
-      );
-      req.user.access_token = access_token;
-      res.send(access_token);
-    }
-  } catch (error) {
-    console.log(error);
-  }
-});
+//       db.collection('users').updateOne(
+//         { _id: req.user._id },
+//         { $set: { access_token } }
+//       );
+//       req.user.access_token = access_token;
+//       res.send(access_token);
+//     }
+//   } catch (error) {
+//     console.log(error);
+//   }
+// });
 
-app.get('/transactions', async (req, res) => {
-  try {
-    const today = new Date();
-    const dd = String(today.getDate()).padStart(2, '0');
-    const mm = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
-    const yyyy = today.getFullYear();
-    // let mmMinusSix = String(today.getMonth() - 1).padStart(2, '0'); //January is 0!
-    // let yyyyFix = today.getFullYear();
+// app.get('/transactions', async (req, res) => {
+//   try {
+//     const today = new Date();
+//     const dd = String(today.getDate()).padStart(2, '0');
+//     const mm = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
+//     const yyyy = today.getFullYear();
 
-    const now = yyyy + '-' + mm + '-' + dd;
-    // let nowMinusSixmm = yyyyFix + '-' + mmMinusSix + '-' + dd;
-    // if (parseInt(mmMinusSix) < 0) {
-    //   mmMinusSix = 12 + parseInt(mmMinusSix);
-    //   yyyyFix--;
-    //   nowMinusSixmm = yyyyFix + '-0' + mmMinusSix + '-' + dd;
-    // }
-    const lastYear = yyyy - 1 + '-' + mm + '-' + dd;
+//     const now = yyyy + '-' + mm + '-' + dd;
+//     const lastYear = yyyy - 1 + '-' + mm + '-' + dd;
 
-    console.log('LAST YEAR', lastYear);
-    const data = await client.getTransactions(
-      req.user.access_token,
-      lastYear,
-      now,
-      {
-        count: 500,
-      }
-    );
+//     console.log('LAST YEAR', lastYear);
+//     const data = await client.getTransactions(
+//       req.user.access_token,
+//       lastYear,
+//       now,
+//       {
+//         count: 500,
+//       }
+//     );
 
-    res.json(data);
-  } catch (error) {
-    console.error(error);
-  }
-});
+//     res.json(data);
+//   } catch (error) {
+//     console.error(error);
+//   }
+// });
 
 app.use(express.static(path.join(__dirname, '../public')));
 app.use('/api', require('./api'));
@@ -166,4 +158,4 @@ app.use(function (err, req, res, next) {
   res.status(err.status || 500).send(err.message || 'Internal server error.');
 });
 
-module.exports = client;
+// module.exports = client;
